requirements.txt
Django>=5.1.6
groq>=0.18.0
python-dotenv>=1.0.1
cartesia>=1.4.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
certifi>=2024.2.2 
psycopg2-binary>=2.9.9
dj-database-url>=2.3.0
whitenoise>=6.0.0
django-filter>=25.1
djangorestframework>=3.15.2
google-search-results>=2.4.2
django-cors-headers>=4.7.0
Pyrebase4==4.8.0
setuptools==80.3.1
google-cloud-storage==3.1.0
openai==1.75.0
firebase-admin>=6.5.0
celery==5.5.3
redis==6.2.0
django-celery-beat==2.8.1